requirements.txt
beautifulsoup4==4.3.2
cssselect==0.9.1
Django==1.9
django-extensions==1.6.1
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
feedfinder2==0.0.1
feedparser==5.1.3
jieba==0.35
lxml==3.3.5
newspaper==0.0.9.8
nltk==2.0.5
Pillow==2.5.1
psycopg2==2.6.1
PyJWT==1.4.0
python-dateutil==2.4.0
PyYAML==3.11
requests==2.3.0
six==1.7.3
tldextract==1.5.1
wheel==0.24.0
